Move tiktoken logic to vendored crate
